Move exeuction code to new folder
