+++ java/execute/src/com/ibm/streamsx/topology/internal/functional/StatelessFunctionalHandler.java
