+++ java/execute/src/com/ibm/streamsx/topology/internal/functional/ops/AbstractPrimitive.java
