+++ java/execute/src/com/ibm/streamsx/topology/internal/functional/ops/Functional.java
