+++ java/execute/src/com/ibm/streamsx/topology/internal/functional/window/KeyPartitioner.java
