+++ java/execute/src/com/ibm/streamsx/topology/internal/functional/window/PeriodicAggregator.java
