+++ java/execute/src/com/ibm/streamsx/topology/internal/spljava/StringMapping.java
