+++ java/execute/src/com/ibm/streamsx/topology/spi/operators/Pipe.java
