+++ java/src/com/ibm/streamsx/topology/spi/TupleSerializer.java
@@ -9,7 +9,6 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.Serializable;
 
-import com.ibm.streamsx.topology.internal.spljava.JavaSerializer;
 
 public interface TupleSerializer extends Serializable {
